Correções e checar suspensaoController

+++ frontend/app/controllers/suspensaoEscalaController.js
@@ -5,6 +5,20 @@
         //Inicializações
         self.motivos = bcUtils.getMotivoSuspensaoEscala();
 
+        self.getAtendimentoAgendandoByProfissional = function(suspensao){
+            $http.get(urls.atendimentos+'/?estado=AGENDADO&profissional='+suspensao.profissional._id).then(function(response){
+                // Filtro pelo período da suspensão
+                self.atendimentoAgendado = response.data.filter(function(atendimento){                      
+                    return new Date(atendimento.dtAtendimento) >= new Date(suspensao.dtInicio) && new Date(atendimento.dtAtendimento) <= new Date(suspensao.dtTermino);
+                }, Object.create(null));
+
+                console.log('Atendimentos na data da suspensão:', self.atendimentoAgendado);
+            }, function(response){
+                console.error('Erro ao buscar atendimentos agendados no período da suspensão: ',  response.data.errors);
+            });
+
+        };
+
         self.saveSuspensao = function(){
             // Validações 
             let dtInicio = new Date(self.suspensao.dtInicio);
@@ -33,25 +47,20 @@
                     }
                 }
             }
+
+            self.getAtendimentoAgendandoByProfissional(self.suspensao);
                 
             $http.post(urls.suspensoes, self.suspensao).then(function(response){
-                // Busca atendimento agendado do profissional no período da suspensão
-                $http.get(urls.atendimentos+'/?estado=AGENDADO&profissional='+self.suspensao.profissional._id).then(function(response){
-                    // Filtro pelo período da suspensão
-                    self.atendimentoAgendado = response.data.filter(function(atendimento){                      
-                        return new Date(atendimento.dtAtendimento) >= new Date(self.suspensao.dtInicio) && new Date(atendimento.dtAtendimento) <= new Date(self.suspensao.dtTermino);
-                    }, Object.create(null));
-
-                    console.log('Atendimentos na data da suspensão:', self.atendimentoAgendado);
-                    //Checa o tipo do dado da resposta e marca o(s) atendimento(s) como cancelado e exclui o(s) agendamento(s)
+                //Checa o tipo do dado da resposta e marca o(s) atendimento(s) como cancelado e exclui o(s) agendamento(s)
+                if(self.atendimentoAgendado){
                     if(self.atendimentoAgendado instanceof Array){
                         for(let i = 0; i < self.atendimentoAgendado.length; i++){
                             self.atendimentoAgendado[i].estado = 'CANCELADO';
                             self.atendimentoAgendado[i].dtCancelamento = new Date();
                             self.atendimentoAgendado[i].descMotivoCancelamento = 'Supensão de escala do profissional';
-
+    
                             $http.put(urls.atendimentos+'/'+self.atendimentoAgendado[i]._id, self.atendimentoAgendado[i]).then(function(response){
-                                console.log('Atendimento ' + self.atendimentoAgendado[i]._id + 'cancelado com sucesso após suspensão de escala!');
+                                console.log('Atendimento ' + self.atendimentoAgendado[i]._id + ' cancelado com sucesso após suspensão de escala!');
                                 $http.delete(urls.escalas+'/'+self.atendimentoAgendado[i].escalaAtendimento).then(function(response){
                                     console.log('Agenda excluída com sucesso após suspensão de escala');                
                                 }, function(response){
@@ -65,9 +74,9 @@
                         self.atendimentoAgendado.estado = 'CANCELADO';
                         self.atendimentoAgendado.dtCancelamento = new Date();
                         self.atendimentoAgendado.descMotivoCancelamento = 'Supensão de escala do profissional';
-
+    
                         $http.put(urls.atendimentos+'/'+self.atendimentoAgendado._id, self.atendimentoAgendado).then(function(response){
-                            console.log('Atendimento ' + self.atendimentoAgendado._id + 'cancelado com sucesso após suspensão de escala!');
+                            console.log('Atendimento ' + self.atendimentoAgendado._id + ' cancelado com sucesso após suspensão de escala!');
                             $http.delete(urls.escalas+'/'+self.atendimentoAgendado.escalaAtendimento).then(function(response){
                                 console.log('Agenda excluída com sucesso após suspensão de escala');                
                             }, function(response){
@@ -76,10 +85,8 @@
                         }, function(response){
                             console.error('Erro ao cancelar atendimento: ', response.data);
                         });
-                    }
-                }, function(response){
-                    console.error('Erro ao buscar atendimentos agendados no período da suspensão: ',  response.data.errors);
-                });
+                    }   
+                }             
                 self.getSuspensoes();
                 msgs.msgSuccess('Suspensão de Escala salva com sucesso!');
             }, function(response){
@@ -89,7 +96,7 @@
         };
 
         self.getSuspensoes = function(){
-            self.atendimentoAgendado = {};
+            self.atendimentoAgendado = undefined;
             $http.get(urls.suspensoes+'?sort=dtInicio').then(function(response){
                 console.log('Atualizando lista de suspensões de escala...');
                 self.suspensao = {};
